# docs/api/openapi.yaml
openapi: 3.1.0
info:
  title: eQuadras API - Gestão e Agendamento Esportivo
  description: "Documentação oficial das APIs REST da plataforma eQuadras para integraç\
    ões de sistemas parceiros, bots e clientes de API. Todos os endpoints incluem\
    \ exemplos reais de requisição e resposta."
  contact:
    name: Suporte Técnico eQuadras
    url: https://github.com/GuilhermeAizzaSano/eQuadras
    email: contato@equadras.app
  license:
    name: MIT License
    url: https://opensource.org/licenses/MIT
  version: v1.0.0
servers:
- url: https://equadras.app
  description: Servidor de Produção (HTTPS)
- url: /
  description: Servidor Relativo (Mesma Origem)
- url: http://localhost:8080
  description: Ambiente Local (Desenvolvimento)
security:
- BearerAuth: []
- ApiKeyAuth: []
...
paths:
  /api/usuarios/{id}:
    get:
      tags:
      - Usuários - API
      summary: Buscar usuário por ID
      description: "Papéis: CLIENT ou ADMIN. O próprio usuário ou o Master Admin.\
        \ Outro usuário recebe 403; ID inexistente, 404 `USUARIO_NAO_ENCONTRADO`."
      operationId: buscarPorId
      parameters:
      - name: id
        in: path
        description: ID do usuário (`id_usuario`)
        required: true
        schema:
          type: integer
          format: int64
        example: 10
      responses:
        "200":
          description: Dados do usuário
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UsuarioResponseDTO"
              example:
                id_usuario: 10
                nome_usuario: Arthur Prado
                email_usuario: arthur.prado@email.com
                phone_usuario: (11) 99999-8888
                role: CLIENT
                criadoEm: 2026-09-04T10:00:00
                masterAdmin: false
        "404":
          description: Recurso Não Encontrado
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/ProblemaErro"
              example:
                type: https://api.equadras.com/erros/recurso-nao-encontrado
                title: Recurso Não Encontrado
                status: 404
                detail: "Usuário não encontrado para o ID: 99"
                instance: "/api/usuarios/{id}"
                code: USUARIO_NAO_ENCONTRADO
                timestamp: 2026-09-29T14:30:00.123456Z
        "403":
          description: Acesso Negado
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/ProblemaErro"
              example:
                type: https://api.equadras.com/erros/acesso-negado
                title: Acesso Negado
                status: 403
                detail: Você não tem permissão para visualizar os dados de outro usuário.
                instance: "/api/usuarios/{id}"
                code: ACESSO_NEGADO
                timestamp: 2026-09-29T14:30:00.123456Z
    put:
      tags:
      - Usuários - API
      summary: Editar usuário existente (Master Admin)
      description: "Papel: ADMIN, exigindo Master Admin. Atualiza nome, e-mail, telefone,\
        \ `role` e, se `nova_senha` vier, a senha. O e-mail do Master Admin não pode\
        \ ser alterado (400 `OPERACAO_NAO_PERMITIDA`)."
      operationId: editar
      parameters:
      - name: id
        in: path
        description: ID do usuário (`id_usuario`)
        required: true
        schema:
          type: integer
          format: int64
        example: 10
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UsuarioEdicaoDTO"
            example:
              nome_usuario: Arthur Prado Silva
              email_usuario: arthur.silva@email.com
              phone_usuario: (11) 99999-9999
              role: CLIENT
              nova_senha: NovaSenha@456
        required: true
      responses:
        "200":
          description: Usuário atualizado
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UsuarioResponseDTO"
              example:
                id_usuario: 10
                nome_usuario: Arthur Prado Silva
                email_usuario: arthur.silva@email.com
                phone_usuario: (11) 99999-9999
                role: CLIENT
                criadoEm: 2026-09-04T10:00:00
                masterAdmin: false
        "422":
          description: Dados Inválidos
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/ProblemaErro"
              example:
                type: https://api.equadras.com/erros/validacao
                title: Dados Inválidos
                status: 422
                detail: "Dados inválidos: Nome: O nome deve ter entre 3 e 80 caracteres"
                instance: "/api/usuarios/{id}"
                code: ERRO_VALIDACAO
                timestamp: 2026-09-29T14:30:00.123456Z
                camposIncorretos:
                - campo: Campo
                  mensagem: Mensagem de validação do campo
        "404":
          description: Recurso Não Encontrado
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/ProblemaErro"
              example:
                type: https://api.equadras.com/erros/recurso-nao-encontrado
                title: Recurso Não Encontrado
                status: 404
                detail: "Usuário não encontrado para o ID: 99"
                instance: "/api/usuarios/{id}"
                code: USUARIO_NAO_ENCONTRADO
                timestamp: 2026-09-29T14:30:00.123456Z
        "400":
          description: Regra de Negócio Violada
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/ProblemaErro"
              examples:
                EMAIL_DUPLICADO:
                  summary: E-mail já cadastrado por outro usuário.
                  value:
                    type: https://api.equadras.com/erros/regra-negocio-violada
                    title: Regra de Negócio Violada
                    status: 400
                    detail: E-mail já cadastrado por outro usuário.
                    instance: "/api/usuarios/{id}"
                    code: EMAIL_DUPLICADO
                    timestamp: 2026-09-29T14:30:00.123456Z
                OPERACAO_NAO_PERMITIDA:
                  summary: O e-mail do Administrador Geral não pode ser modificado.
                  value:
                    type: https://api.equadras.com/erros/regra-negocio-violada
                    title: Regra de Negócio Violada
                    status: 400
                    detail: O e-mail do Administrador Geral não pode ser modificado.
                    instance: "/api/usuarios/{id}"
                    code: OPERACAO_NAO_PERMITIDA
                    timestamp: 2026-09-29T14:30:00.123456Z
        "403":
          description: Acesso Negado
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/ProblemaErro"
              example:
                type: https://api.equadras.com/erros/acesso-negado
                title: Acesso Negado
                status: 403
                detail: Acesso restrito ao Administrador Geral do sistema.
                instance: "/api/usuarios/{id}"
                code: ACESSO_NEGADO
                timestamp: 2026-09-29T14:30:00.123456Z
    delete:
      tags:
      - Usuários - API
      summary: Excluir usuário (Master Admin)
      description: "Papel: ADMIN, exigindo Master Admin. Remove o usuário. A conta\
        \ do Master Admin não pode ser excluída (400 `OPERACAO_NAO_PERMITIDA`)."
      operationId: excluir
      parameters:
      - name: id
        in: path
        description: ID do usuário (`id_usuario`)
        required: true
        schema:
          type: integer
          format: int64
        example: 10
      responses:
        "204":
          description: Usuário excluído
        "404":
          description: Recurso Não Encontrado
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/ProblemaErro"
              example:
                type: https://api.equadras.com/erros/recurso-nao-encontrado
                title: Recurso Não Encontrado
                status: 404
                detail: "Usuário não encontrado para o ID: 99"
                instance: "/api/usuarios/{id}"
                code: USUARIO_NAO_ENCONTRADO
                timestamp: 2026-09-29T14:30:00.123456Z
        "400":
          description: Regra de Negócio Violada
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/ProblemaErro"
              example:
                type: https://api.equadras.com/erros/regra-negocio-violada
                title: Regra de Negócio Violada
                status: 400
                detail: A conta do Administrador Geral não pode ser excluída.
                instance: "/api/usuarios/{id}"
                code: OPERACAO_NAO_PERMITIDA
                timestamp: 2026-09-29T14:30:00.123456Z
        "403":
          description: Acesso Negado
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/ProblemaErro"
              example:
                type: https://api.equadras.com/erros/acesso-negado
                title: Acesso Negado
                status: 403
                detail: Acesso restrito ao Administrador Geral do sistema.
                instance: "/api/usuarios/{id}"
                code: ACESSO_NEGADO
                timestamp: 2026-09-29T14:30:00.123456Z
...
  /api/notificacoes/{id}/ler:
    put:
      tags:
      - Notificações - API
      summary: Marcar notificação como lida
      description: "Papel: ADMIN. Marca uma notificação do próprio administrador como\
        \ lida. Notificação de outro usuário devolve 400. Responde 204 sem corpo."
      operationId: marcarComoLida
      parameters:
      - name: id
        in: path
        description: ID da notificação
        required: true
        schema:
          type: integer
          format: int64
        example: 1
      responses:
        "204":
          description: Notificação marcada como lida
        "400":
          description: Requisição Inválida
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/ProblemaErro"
              example:
                type: https://api.equadras.com/erros/bad-request
                title: Requisição Inválida
                status: 400
                detail: Você não tem permissão para alterar esta notificação.
                instance: "/api/notificacoes/{id}/ler"
                code: REQUISICAO_INVALIDA
                timestamp: 2026-09-29T14:30:00.123456Z
        "403":
          description: Acesso Proibido
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/ProblemaErro"
              example:
                status: 403
                title: Acesso Proibido
                detail: "Acesso proibido: sua credencial não possui permissão para\
                  \ executar esta operação."
...
  /api/usuarios/login:
    post:
      tags:
      - Usuários - API
      summary: Realizar login
      description: "Rota pública. Autentica por e-mail e senha, grava o JWT no cookie\
        \ HttpOnly `equadras_session` (header `Set-Cookie`) e devolve no corpo apenas\
        \ o perfil do usuário (`UsuarioResponseDTO`); o token NÃO vem no corpo. Efeito\
        \ colateral: falhas consecutivas por e-mail (limite de 5) bloqueiam novas\
        \ tentativas com 429 e `Retry-After`; login bem-sucedido zera o contador."
      operationId: login
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UsuarioLoginDTO"
            example:
              email_usuario: arthur.prado@email.com
              senha_usuario: SenhaSegura@123
        required: true
      responses:
        "200":
          description: Login realizado. O corpo traz o perfil; o JWT vai apenas no
            cookie HttpOnly `equadras_session`.
          headers:
            Set-Cookie:
              description: "Cookie de sessão `equadras_session` (HttpOnly, SameSite=Lax,\
                \ Path=/)."
              schema:
                type: string
          content:
            application/json:
              schema:
                type: object
              example:
                id_usuario: 10
                nome_usuario: Arthur Prado
                email_usuario: arthur.prado@email.com
                phone_usuario: (11) 99999-8888
                role: CLIENT
                criadoEm: 2026-09-04T10:00:00
                masterAdmin: false
        "422":
          description: Dados Inválidos
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/ProblemaErro"
              example:
                type: https://api.equadras.com/erros/validacao
                title: Dados Inválidos
                status: 422
                detail: "Dados inválidos: E-mail: Formato de e-mail inválido"
                instance: /api/usuarios/login
                code: ERRO_VALIDACAO
                timestamp: 2026-09-29T14:30:00.123456Z
                camposIncorretos:
                - campo: Campo
                  mensagem: Mensagem de validação do campo
        "400":
          description: Regra de Negócio Violada
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/ProblemaErro"
              example:
                type: https://api.equadras.com/erros/regra-negocio-violada
                title: Regra de Negócio Violada
                status: 400
                detail: E-mail ou senha incorretos.
                instance: /api/usuarios/login
                code: CREDENCIAIS_INVALIDAS
                timestamp: 2026-09-29T14:30:00.123456Z
        "429":
          description: Too Many Requests
          headers:
            Retry-After:
              description: Segundos até nova tentativa de login para este e-mail.
              schema:
                type: string
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/ProblemaErro"
              example:
                type: about:blank
                title: Too Many Requests
                status: 429
                detail: Muitas tentativas falhas de login para esta conta. Tente novamente
                  em 300 segundos.
      security: []
  /api/usuarios/api-key/regenerar:
    post:
      tags:
      - Usuários - API
      summary: Gerar ou regenerar API-KEY
      description: "Papéis: CLIENT ou ADMIN. Emite uma chave opaca `eq_...`, invalidando\
        \ a anterior imediatamente. A chave em texto plano só aparece nesta resposta\
        \ (`Cache-Control: no-store`). Registra IP e User-Agent. Limite: 5 regeneraç\
        ões por minuto por conta (429 com `Retry-After: 60`)."
      operationId: regenerarApiKey
      responses:
        "200":
          description: Nova chave emitida. A chave em texto plano só aparece nesta
            resposta.
          headers:
            Cache-Control:
              description: "`no-store, no-cache, must-revalidate, max-age=0`: a resposta\
                \ não pode ser guardada em cache."
              schema:
                type: string
          content:
            application/json:
              schema:
                type: object
              example:
                apiKey: eq_9f8e7d6c5b4a39281706f5e4d3c2b1a0
                last4: b1a0
                criadaEm: 2026-09-29T14:30:00Z
        "401":
          description: Não Autorizado
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/ProblemaErro"
              example:
                status: 401
                title: Não Autorizado
                detail: "Acesso não autorizado: credencial ausente, expirada ou invá\
                  lida."
        "429":
          description: Too Many Requests
          headers:
            Retry-After:
              description: Sempre 60 (segundos).
              schema:
                type: string
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/ProblemaErro"
              example:
                status: 429
                title: Too Many Requests
                detail: Limite de 5 regenerações por minuto atingido para esta conta.
                  Aguarde antes de tentar novamente.
...
  /api/usuarios/me:
    get:
      tags:
      - Usuários - API
      summary: Dados da minha sessão
      description: "Papéis: CLIENT ou ADMIN. Devolve o perfil do usuário autenticado\
        \ (cookie de sessão, Bearer ou X-API-KEY)."
      operationId: me
      responses:
        "200":
          description: Perfil do usuário autenticado
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UsuarioResponseDTO"
              example:
                id_usuario: 10
                nome_usuario: Arthur Prado
                email_usuario: arthur.prado@email.com
                phone_usuario: (11) 99999-8888
                role: CLIENT
                criadoEm: 2026-09-04T10:00:00
                masterAdmin: false
        "401":
          description: Não Autorizado
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/ProblemaErro"
              example:
                status: 401
                title: Não Autorizado
                detail: "Acesso não autorizado: credencial ausente, expirada ou invá\
                  lida."
...
components:
  schemas:
...
    UsuarioResponseDTO:
      type: object
      description: Perfil cadastral de um usuário do sistema
      properties:
        criadoEm:
          type: string
          format: date-time
          description: Data e hora de criação da conta
          example: 2026-09-04T10:00:00
        email_usuario:
          type: string
          description: Endereço de e-mail
          example: arthur.prado@email.com
        id_usuario:
          type: integer
          format: int64
          description: ID do usuário
          example: 10
        masterAdmin:
          type: boolean
          description: Indica se o usuário possui privilégios de Master Admin
          example: false
        nome_usuario:
          type: string
          description: Nome completo
          example: Arthur Prado
        phone_usuario:
          type: string
          description: Telefone ou WhatsApp
          example: (11) 99999-8888
        role:
          type: string
          description: Papel no sistema (CLIENT ou ADMIN)
          enum:
          - CLIENT
          - ADMIN
          example: CLIENT
...
    UsuarioLoginDTO:
      type: object
      properties:
        email_usuario:
          type: string
          description: E-mail de acesso da conta
          example: arthur.prado@email.com
        senha_usuario:
          type: string
          description: Senha da conta
          example: SenhaSegura@123
      required:
      - email_usuario
      - senha_usuario
...
    ProblemaErro:
      type: object
      description: "Corpo de erro (RFC 9457). Os filtros de segurança devolvem só\
        \ status, title e detail."
      properties:
        type:
          type: string
        title:
          type: string
        status:
          type: integer
          format: int32
        detail:
          type: string
        instance:
          type: string
        code:
          type: string
        timestamp:
          type: string
        camposIncorretos:
          type: array
          items:
            type: object
            properties:
              campo:
                type: string
              mensagem:
                type: string
...
  securitySchemes:
    BearerAuth:
      type: http
      description: "Informe o token JWT ou API-Key no formato: `Bearer eq_...` gerado\
        \ no painel da conta."
      scheme: bearer
      bearerFormat: JWT
    ApiKeyAuth:
      type: apiKey
      description: Chave de API pessoal gerada no painel do usuário (formato `eq_...`).
        Válida em todas as rotas da API respeitando os papéis (ROLE_CLIENT e ROLE_ADMIN).
      name: X-API-KEY
      in: header